Enhance leave management functionality and data types
- Added new using directives and dependencies in LeaveRequestManager.
- Refactored CreateLeaveRequestAsync for better error handling and logic.
- Updated LeaveConstants to include TwoHourExcuse and related constants.
- Introduced new properties in LeaveRequest for hourly leave tracking.
- Changed UsedDays type in LeaveBalance from int to decimal for precision.
- Updated KmtDbContext to reflect new database schema changes.
- Created migration files for database adjustments.
- Introduced LeaveRequestStatus enum for request statuses.
- Modified UpdateLeaveBalanceRequest to align UsedDays type with LeaveBalance.

+++ KmtBackend.DAL/Entities/LeaveBalance.cs
@@ -27,10 +27,10 @@ namespace KmtBackend.DAL.Entities
 
         // Days already used
         [Required]
-        public int UsedDays { get; set; }
+        public decimal UsedDays { get; set; }
 
         // Computed property for remaining days
         [NotMapped]
-        public int RemainingDays => TotalDays - UsedDays;
+        public decimal RemainingDays => TotalDays - UsedDays;
     }
 }
